Factor out encrypted_body middleware and endpoint into a "sub app"
to avoid using a global BodyDecryptor

This required one of the handlers tests to change because there is
no longer a global available. It now better represents what a client
would do anyway.
