+++ src/handlers.js
@@ -23,7 +23,6 @@ const validate = require('express-validation');
 
 const { getReport, generateReportFromDownload } = require('./reporting.js');
 const withTempDir = require('./with-temp-dir.js');
-const BodyDecryptor = require('./decrypt-body.js');
 
 const ClientError = require('./client-error.js');
 
@@ -200,15 +199,6 @@ async function scanReportHandler(req, res, next, matrixFile) {
     res.status(200).json(responseBody);
 }
 
-function publicKeyHandler(req, res, next) {
-    const decryptor = BodyDecryptor.getDecryptor();
-    const responseBody = {
-        public_key: decryptor.getPublicKey(),
-    };
-
-    res.status(200).json(responseBody);
-}
-
 function attachHandlers(app) {
     app.post(
         '/_matrix/media_proxy/unstable/download_encrypted',
@@ -235,7 +225,6 @@ function attachHandlers(app) {
         validate(unencryptedRequestSchema),
         wrapAsyncHandle(scanReportHandler)
     );
-    app.get('/_matrix/media_proxy/unstable/public_key', publicKeyHandler);
 }
 
 module.exports = {
