+++ test/handlers-test.js
@@ -24,8 +24,6 @@ const { clearReportCache } = require('../src/reporting.js');
 const app = require('../src/app.js').createApp();
 const example = require('../example.file.json');
 
-const BodyDecryptor = require('../src/decrypt-body.js');
-
 const { setConfig } = require('../src/config.js');
 
 setConfig({
@@ -95,8 +93,10 @@ describe('POST /_matrix/media_proxy/unstable/scan_encrypted', () => {
     it('responds with a scan report if `encrypted_body` is given', () => {
         const plainBody = { file: example.file };
 
-        const decryptor = BodyDecryptor.getDecryptor();
-        const publicKey = decryptor.getPublicKey();
+        const publicKey = await request(app)
+            .get('/_matrix/media_proxy/unstable/public_key')
+            .then(response => response.body.public_key);
+
         const encryption = new PkEncryption();
         encryption.set_recipient_key(publicKey);
         const encryptedBody = encryption.encrypt(JSON.stringify(plainBody));
